--- src/src/App.css
@@ -1,28 +1,0 @@
-.application {
-  text-align: center;
-}
-
-.logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 80px;
-}
-
-.header {
-  background-color: #222;
-  height: 150px;
-  padding: 20px;
-  color: white;
-}
-
-.title {
-  font-size: 1.5em;
-}
-
-.intro {
-  font-size: large;
-}
-
-@keyframes App-logo-spin {
-  from { transform: rotate(0deg); }
-  to { transform: rotate(360deg); }
-}
