--- src/src/assets/index.js
@@ -1,21 +1,0 @@
-import image0 from './lego0.jpg';
-import image1 from './lego1.jpg';
-import image2 from './lego2.jpg';
-import image3 from './lego3.jpg';
-import image4 from './lego4.jpg';
-import image5 from './lego5.jpg';
-import image6 from './lego6.jpg';
-import image7 from './lego7.jpg';
-import image8 from './lego8.jpg';
-
-export {
-    image0,
-    image1,
-    image2,
-    image3,
-    image4,
-    image5,
-    image6,
-    image7,
-    image8,
-};
