+++ src/utils/fixtures.js
@@ -1,0 +1,64 @@
+import * as images from '../assets';
+
+const fixtures = {
+    "users": [
+        {
+            name: "Tony",
+            netflix: true,
+            amazon: false,
+            hulu: false,
+            image: images.image0,
+            bio: "Snare ultimate eternal-return pious merciful justice convictions sea superiority free. Sexuality philosophy love derive burying enlightenment marvelous.",
+        },
+        {
+            name: "Bruce",
+            netflix: true,
+            amazon: true,
+            hulu: true,
+            image: images.image1,
+            bio: "Oneself ideal against ubermensch strong mountains gains transvaluation eternal-return convictions merciful decrepit endless selfish. Free will faith.",
+        },
+        {
+            name: "Steve",
+            netflix: false,
+            amazon: true,
+            hulu: true,
+            image: images.image2,
+            bio: "Christianity oneself abstract christian ultimate ultimate hatred christianity decieve right fearful virtues truth. Morality mountains intentions.",
+        },
+        {
+            name: "Nick",
+            netflix: false,
+            amazon: false,
+            hulu: true,
+            image: images.image3,
+            bio: "Play joy enlightenment hatred sea ultimate pious dead ocean overcome. Morality aversion decrepit strong morality burying ascetic decieve.",
+        },
+        {
+            name: "Peter",
+            netflix: false,
+            amazon: true,
+            hulu: false,
+            image: images.image4,
+            bio: "Will philosophy convictions oneself inexpedient justice joy sea transvaluation holiest suicide marvelous contradict. Depths endless.",
+        },
+        {
+            name: "Bucky",
+            netflix: true,
+            amazon: false,
+            hulu: true,
+            image: images.image5,
+            bio: "Faithful truth mountains fearful moral joy love reason aversion moral selfish prejudice. Sexuality endless gains.",
+        },
+        {
+            name: "Logan",
+            netflix: false,
+            amazon: true,
+            hulu: true,
+            image: images.image6,
+            bio: "Snare evil enlightenment against snare suicide dead horror deceptions decieve burying victorious truth. Of contradict.",
+        },
+    ],
+};
+
+export default fixtures;
